Conflicts:
	arangod/Indexes/HashIndex.cpp

+++ UnitTests/Basics/AttributeNameParserTest.cpp
@@ -68,7 +68,7 @@ BOOST_AUTO_TEST_CASE (test_simpleString) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 1);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(1));
   BOOST_CHECK_EQUAL(result[0].name, input);
   BOOST_CHECK_EQUAL(result[0].shouldExpand, false);
 }
@@ -83,7 +83,7 @@ BOOST_AUTO_TEST_CASE (test_subAttribute) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 2);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(2));
   BOOST_CHECK_EQUAL(result[0].name, "foo");
   BOOST_CHECK_EQUAL(result[0].shouldExpand, false);
   BOOST_CHECK_EQUAL(result[1].name, "bar");
@@ -100,7 +100,7 @@ BOOST_AUTO_TEST_CASE (test_subsubAttribute) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 3);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(3));
   BOOST_CHECK_EQUAL(result[0].name, "foo");
   BOOST_CHECK_EQUAL(result[0].shouldExpand, false);
   BOOST_CHECK_EQUAL(result[1].name, "bar");
@@ -119,7 +119,7 @@ BOOST_AUTO_TEST_CASE (test_expandAttribute) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 1);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(1));
   BOOST_CHECK_EQUAL(result[0].name, "foo");
   BOOST_CHECK_EQUAL(result[0].shouldExpand, true);
 }
@@ -134,7 +134,7 @@ BOOST_AUTO_TEST_CASE (test_expandSubAttribute) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 2);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(2));
   BOOST_CHECK_EQUAL(result[0].name, "foo");
   BOOST_CHECK_EQUAL(result[0].shouldExpand, false);
   BOOST_CHECK_EQUAL(result[1].name, "bar");
@@ -151,7 +151,7 @@ BOOST_AUTO_TEST_CASE (test_expandedSubAttribute) {
 
   TRI_ParseAttributeString(input, result);
   
-  BOOST_CHECK_EQUAL(result.size(), 2);
+  BOOST_CHECK_EQUAL(result.size(), static_cast<size_t>(2));
   BOOST_CHECK_EQUAL(result[0].name, "foo");
   BOOST_CHECK_EQUAL(result[0].shouldExpand, true);
   BOOST_CHECK_EQUAL(result[1].name, "bar");
